requirements.txt
camelot-py
langchain-openai
langchain-community
unstructured
PyMuPDF
networkx
omegaconf
opencv-python